refactor/#29: 회원가입 하위 페이지 components 디렉토리로 이동
